backend/auth_server/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
requests
bcrypt
pyjwt
python-dotenv
